+++ src/OpenTelemetry.Instrumentation.AWS/Implementation/AWSServiceHelper.cs
@@ -23,12 +23,13 @@ internal class AWSServiceHelper
                 .AddAttributeAWSBedrockKnowledgeBaseId("KnowledgeBaseId")
                 .AddAttributeAWSS3BucketName("BucketName")
                 .AddAttributeAWSS3Key("Key")
+                .AddAttributeMessageId("MessageId")
                 .Build();
 
         this.ArrayValueAttributeNames = new(semanticConventions.GetArrayValueAttributeNames(), StringComparer.Ordinal);
     }
 
-    internal static IReadOnlyDictionary<string, List<string>> ServiceRequestParameterMap { get; } = new Dictionary<string, List<string>>()
+    internal static IReadOnlyDictionary<string, List<string>> ServiceRequestParameterMap { get; } = new Dictionary<string, List<string>>(7)
     {
         { AWSServiceType.DynamoDbService, ["TableName"] },
         { AWSServiceType.S3Service, ["BucketName", "Key"] },
@@ -39,10 +40,28 @@ internal class AWSServiceHelper
         { AWSServiceType.BedrockRuntimeService, ["ModelId"] },
     };
 
-    internal static IReadOnlyDictionary<string, List<string>> ServiceResponseParameterMap { get; } = new Dictionary<string, List<string>>()
+    internal static IReadOnlyDictionary<string, List<string>> ServiceResponseParameterMap { get; } = new Dictionary<string, List<string>>(4)
     {
         { AWSServiceType.BedrockService, ["GuardrailId"] },
         { AWSServiceType.BedrockAgentService, ["AgentId", "DataSourceId"] },
+        { AWSServiceType.SNSService, ["MessageId", "TopicArn"] },
+        { AWSServiceType.SQSService, ["MessageId", "QueueUrl"] },
+    };
+
+    internal static IReadOnlyDictionary<string, IReadOnlyDictionary<string, string>> MessagingOperationTypeMap { get; } = new Dictionary<string, IReadOnlyDictionary<string, string>>(2)
+    {
+        // See https://github.com/open-telemetry/semantic-conventions/blob/v1.40.0/docs/registry/attributes/messaging.md#messaging-operation-type
+        [AWSServiceType.SNSService] = new Dictionary<string, string>(2)
+        {
+            ["Publish"] = "send",
+            ["PublishBatch"] = "send",
+        },
+        [AWSServiceType.SQSService] = new Dictionary<string, string>(3)
+        {
+            ["ReceiveMessage"] = "receive",
+            ["SendMessage"] = "send",
+            ["SendMessageBatch"] = "send",
+        },
     };
 
     // for Bedrock Agent operations, we map each supported operation to one resource: Agent, DataSource, or KnowledgeBase