+++ src/OpenTelemetry.Instrumentation.AWS/CHANGELOG.md
@@ -14,6 +14,9 @@
 * Add `cloud.region` attribute to all AWS SDK client spans.
   ([#4043](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4043))
 
+* Add messaging attributes for AWS SNS and SQS.
+  ([#4043](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4043))
+
 * BREAKING: Update latest AWS Semantic Conventions to 1.40.0.
   ([#4043](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4043))